721. PCI chip ID updates to ATI driver (Marc La France).  720. Fix i810 driver for -probe and -configure (Marc La France).  719. Change message when default modes are deleted (Marc La France).  718. Fix Xinerama byte swapping bug (Marc La France).  717. IA-64 and Alpha fixes for pswrap, Mesa, DRI, Xpm, libX11, Xt, Xaw, Xmu,       dps, Type1 fonts, cfb24, most output drivers, ELF loader, ramdac module,       xf4bpp and xf86cfg (Marc La France).  716. Improve IA-64 support by removing a plethora of 32-bit'isms       (Marc La France).  715. Default HasLinuxDoc to NO (Marc La France).
+ ident lines and warning fixes.

+++ lib/X11/imRm.c
@@ -30,7 +30,7 @@ PERFORMANCE OF THIS SOFTWARE.
 			       makoto@sm.sony.co.jp
 
 ******************************************************************/
-/* $XFree86: xc/lib/X11/imRm.c,v 3.7 1999/05/09 10:50:36 dawes Exp $ */
+/* $XFree86: xc/lib/X11/imRm.c,v 3.8 2000/01/29 18:58:15 dawes Exp $ */
 
 #include <stdio.h>
 #include <X11/Xlib.h>
@@ -953,7 +953,7 @@ _XimEncodeLineSpace(info, top, val)
     int			*out;
 
     out = (int *)((char *)top + info->offset);
-    *out = (int)val;
+    *out = (long)val;
     return True;
 }
